[#177396474] Allow configurable max age for CGN and EYCA
