+++ StartCgnActivation/index.ts
@@ -38,7 +38,7 @@ secureExpressApp(app);
 // Add express route
 app.post(
   "/api/v1/cgn/:fiscalcode/activation",
-  StartCgnActivation(userCgnModel)
+  StartCgnActivation(userCgnModel, config.CGN_BETA_TEST_UPPER_BOUND_AGE)
 );
 
 const azureFunctionHandler = createAzureFunctionHandler(app);
