+++ UpdateExpiredCgn/handler.ts
@@ -9,7 +9,7 @@ import { NonEmptyString } from "italia-ts-commons/lib/strings";
 import { StatusEnum as CardActivatedStatusEnum } from "../generated/definitions/CardActivated";
 import { StatusEnum as CardExpiredStatusEnum } from "../generated/definitions/CardExpired";
 import { StatusEnum as CardRevokedStatusEnum } from "../generated/definitions/CardRevoked";
-import { OrchestratorInput } from "../UpdateCgnOrchestrator";
+import { OrchestratorInput } from "../UpdateCgnOrchestrator/handler";
 import { initTelemetryClient, trackException } from "../utils/appinsights";
 import {
   makeUpdateCgnOrchestratorId,
